requirements.txt
Flask
python-dotenv
langchain
langchain-google-genai
